ДЗ7. Переписано под Spring Boot, добавлен шедулер.
