+++ build.gradle
@@ -1,5 +1,7 @@
 plugins {
     id 'java'
+    id 'org.springframework.boot' version '2.7.18'
+    id 'io.spring.dependency-management' version '1.1.4'
 }
 
 group = 'ru.mts'
@@ -12,12 +14,17 @@ repositories {
 dependencies {
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter:5.10.0'
-    implementation 'org.springframework:spring-core:6.1.3'
-    implementation 'org.springframework:spring-context:6.1.3'
-    implementation 'org.springframework:spring-beans:6.1.3'
     implementation 'javax.annotation:javax.annotation-api:1.3.2'
+    implementation 'org.springframework.boot:spring-boot-starter'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    implementation 'org.quartz-scheduler:quartz:2.3.2'
+    compileOnly 'org.projectlombok:lombok:1.18.30'
 }
 
 test {
     useJUnitPlatform()
+}
+
+java {
+    sourceCompatibility = '11'
 }
