--- src/main/java/ru/mts/hw_6/config/CreateAnimalBeanPostProcessor.java
@@ -1,23 +1,0 @@
-package ru.mts.hw_6.config;
-
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.config.BeanPostProcessor;
-import org.springframework.lang.Nullable;
-import ru.mts.hw_6.service.CreateAnimalService;
-
-public class CreateAnimalBeanPostProcessor implements BeanPostProcessor {
-    @Nullable
-//    @Override
-    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
-        return bean;
-    }
-
-    @Nullable
-//    @Override
-    public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
-        if (bean instanceof CreateAnimalService) {
-            ((CreateAnimalService) bean).initAnimal();
-        }
-        return bean;
-    }
-}
