--- src/main/java/ru/mts/hw_6/factory/AnimalTypes.java
@@ -1,8 +1,0 @@
-package ru.mts.hw_6.factory;
-
-public enum AnimalTypes {
-    WOLF,
-    SHARK,
-    RABBIT,
-    DEER
-}
