+++ src/main/java/ru/mts/hw_7/factory/ConcreteAnimalFactory.java
@@ -1,0 +1,10 @@
+package ru.mts.hw_7.factory;
+
+import ru.mts.hw_7.animal.Animal;
+
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+public interface ConcreteAnimalFactory {
+    Animal createAnimal(String breed, BigDecimal cost, LocalDate birthDate);
+}
