+++ src/main/resources/application.properties
@@ -1,0 +1,1 @@
+animal.wolf.names=Antoshka,Oleja,Deniska,Vladno
